nvme: Migrate definition of NVMeDrivePresence class
Previously this was hidden inside the Everest support for the Basecamp
board. Move it to the NVMe header as it can also be used for Bonnell.
