+++ src/platforms/everest/basecamp.cpp
@@ -78,39 +78,6 @@ SysfsI2CBus Basecamp::getDriveBus(int index)
     return {driveMux, driveChannelMap.at(index)};
 }
 
-class BasecampNVMeDrivePresence
-{
-  public:
-    BasecampNVMeDrivePresence() = delete;
-    BasecampNVMeDrivePresence(gpiod::line&& line, SysfsI2CBus bus) :
-        line(line), bus(std::move(bus)), haveEndpoint(false)
-    {}
-
-    bool operator()()
-    {
-        // Once we've observed both GPIO presence and the basic I2C endpoint,
-        // only unplug() the drive if the GPIO indicates the drive is unplugged.
-        // The I2C endpoint will come and go as the host power state changes.
-        if (line.get_value() == 0)
-        {
-            haveEndpoint = false;
-            return false;
-        }
-
-        if (!haveEndpoint)
-        {
-            haveEndpoint = BasicNVMeDrive::isBasicEndpointPresent(bus);
-        }
-
-        return haveEndpoint;
-    }
-
-  private:
-    gpiod::line line;
-    SysfsI2CBus bus;
-    bool haveEndpoint;
-};
-
 void Basecamp::plug(Notifier& notifier)
 {
     SysfsI2CBus root(driveMetadataBus);
@@ -128,8 +95,7 @@ void Basecamp::plug(Notifier& notifier)
         auto line = chip.get_line(drivePresenceMap[index]);
         line.request({program_invocation_short_name,
                       gpiod::line::DIRECTION_INPUT, gpiod::line::ACTIVE_LOW});
-        auto presence = BasecampNVMeDrivePresence(std::move(line),
-                                                  getDriveBus(index));
+        auto presence = NVMeDrivePresence(std::move(line), getDriveBus(index));
         poller.start(notifier, std::move(presence));
     }
 }
